+++ src/ace/ext-browse_ss.d.ts
@@ -3,6 +3,7 @@ declare global {
     interface Window {
         _browseSs_DEBUG: Boolean?,
         _browseSsDebugLog: Function,
+        _browseSsLastPrompt: { popup: any, cmdLine: any },
         appDMS: any,
         _browseSsFilesListener: KeyboardEventListener;
         _browseSsLibraryListener: KeyboardEventListener;